SJIP-209: Refactor QueryBuilder State Management
