+++ packages/ui/Release.md
@@ -1,5 +1,9 @@
+### 4.0.0 | 2022-04-01
+- Breaking Changes / feat: Refactor QueryBuilder State Management
+
 ### 3.3.9 | 2022-03-31
 - fix: Build erreor With sidebarMenu
+
 ### 3.3.5 | 2022-03-28
 - feat: Add facet name as ReactNode.
