+++ packages/ui/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@ferlab/ui",
-    "version": "3.3.9",
+    "version": "4.0.0",
     "description": "Core components for scientific research data portals",
     "publishConfig": {
         "access": "public"
